Complete re-organization of directories. Also MonoMorphisized the PlayerTrait shifting / square manipulation functions.
