+++ src/board/eval.rs
@@ -2,9 +2,10 @@
 
 use board::Board;
 use std::i16;
-use templates::*;
-use bit_twiddles::*;
-use templates::{PlayerTrait};
+use core::templates::*;
+use core::bit_twiddles::*;
+use core::templates::{PlayerTrait};
+use core::masks::*;
 
 lazy_static! {
     pub static ref BISHOP_POS: [[i16; SQ_CNT]; PLAYER_CNT] = [ flatten(flip(BISHOP_POS_ARRAY)), flatten(BISHOP_POS_ARRAY) ];
