+++ src/core/mod.rs
@@ -1,0 +1,5 @@
+pub mod bit_twiddles;
+pub mod piece_move;
+pub mod templates;
+pub mod magic_helper;
+pub mod masks;
