+++ src/engine.rs
@@ -2,10 +2,10 @@
 
 extern crate rand;
 
-use piece_move::BitMove;
+use core::piece_move::BitMove;
 use timer::Timer;
 use board::Board;
-use templates::Player;
+use core::templates::Player;
 
 use std::sync::Arc;
 use std::sync::atomic::AtomicBool;
