+++ tests/board_move_apply.rs
@@ -2,7 +2,7 @@ extern crate pleco;
 extern crate rand;
 
 use pleco::board::Board;
-use pleco::piece_move::BitMove;
+use pleco::core::piece_move::BitMove;
 use pleco::tools::gen_rand_legal_board;
 use std::*;
